+++ src/main/java/pokeraidbot/commands/WhatsNewCommand.java
@@ -25,12 +25,18 @@ public class WhatsNewCommand extends ConfigAwareCommand {
         final String message;
         if (config.getLocale().equals(LocaleService.SWEDISH)) {
             message = "**Nytt i 1.8.0-" + BotServerMain.version + ":**\n\n" +
+                    "* Ny logik för grupper som skapas automatiskt via botintegration, " +
+                    "om det finns mer än en aktiv tier5-boss, " +
+                    "sätt gruppens starttid till 30 min efter kläckning istället för 10 så folk hinner se vad det blev\n" +
                     "* Nya raidbossar\n" +
                     "* Gym data uppdaterad\n" +
                     "* Möjligt att markera gymnamn om de är potentiella EX gyms (kräver viss extradata)\n" +
                     "* Celebi och Mew preppade som nya bossar\n";
         } else {
             message = "**New in 1.8.0-" + BotServerMain.version + ":**\n\n" +
+                    "* New handling of automatically created raid groups via bot integration, " +
+                    "if more than one tier 5 boss is active, " +
+                    "set start time to 30 min after hatch instead of 10, so people can find out what boss it was\n" +
                     "* New raidbosses\n" +
                     "* Gym data updated\n" +
                     "* Possible to mark gymnames if they are potential EX raid gyms (requires extra data)\n" +